requirements.txt
absl-py==2.4.0
asgiref==3.11.1
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
cloudinary==1.44.1
Django==5.2.12
django-filter==25.2
djangorestframework==3.17.1
filelock==3.25.2
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.2
grpcio==1.80.0
h5py==3.14.0
idna==3.11
joblib==1.5.3
keras==3.12.1
libclang==18.1.1
markdown-it-py==4.0.0
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.2.6
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
playwright==1.58.0
protobuf==7.34.1
psycopg2-binary==2.9.11
pyee==13.0.1
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2026.1.post1
requests==2.33.1
requests-file==3.0.1
rich==14.3.3
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sqlparse==0.5.5
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
tldextract==5.3.1
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wrapt==2.1.2
